requirements.txt
flask==2.3.3
flask-wtf==1.1.1
flask-login==0.6.2
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
psutil==5.9.5
Flask-APScheduler==1.12.4